rft: Rearrange responsibility
  - Revoke responsibility(register, update) from MockTimeSource, give it to TimeSource
  - Pull members up from MockTimeSource to TimeSource, Responsibility get more clear, and remove duplicate calls
  - Change TimeSource to Class (not Interface anymore)
